+++ src/FilamentApexChartsServiceProvider.php
@@ -2,20 +2,15 @@
 
 namespace Leandrocfe\FilamentApexCharts;
 
-use Filament\Support\Assets\Js;
+use Filament\Support\Assets\AlpineComponent;
 use Filament\Support\Facades\FilamentAsset;
 use Illuminate\Support\Facades\Blade;
 use Leandrocfe\FilamentApexCharts\Commands\FilamentApexChartsCommand;
-use Leandrocfe\FilamentApexCharts\Widgets\ApexChartWidget;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
 
 class FilamentApexChartsServiceProvider extends PackageServiceProvider
 {
-    protected array $widgets = [
-        ApexChartWidget::class,
-    ];
-
     /**
      * Configures the given package with the name 'filament-apex-charts'
      * as a Package Service Provider.
@@ -43,15 +38,13 @@ class FilamentApexChartsServiceProvider extends PackageServiceProvider
      *
      * @throws \Exception If the component namespace is not valid.
      */
-    public function bootingPackage()
+    public function packageBooted(): void
     {
+        parent::packageBooted();
         Blade::componentNamespace('Leandrocfe\\FilamentApexCharts\\Components', 'filament-apex-charts');
-    }
 
-    public function packageRegistered(): void
-    {
         FilamentAsset::register([
-            Js::make('apexcharts', __DIR__.'/../dist/apexcharts.js'),
-        ], package: 'leandrocfe/filament-apex-charts');
+            AlpineComponent::make('apexcharts', __DIR__.'/../dist/apexcharts.js'),
+        ]);
     }
 }