+++ src/Widgets/ApexChartWidget.php
@@ -3,6 +3,7 @@
 namespace Leandrocfe\FilamentApexCharts\Widgets;
 
 use Filament\Forms\Contracts\HasForms;
+use Filament\Support\RawJs;
 use Filament\Widgets\Concerns\CanPoll;
 use Filament\Widgets\Widget;
 use Illuminate\Contracts\View\View;
@@ -17,14 +18,14 @@ use Leandrocfe\FilamentApexCharts\Concerns\HasLoadingIndicator;
 
 class ApexChartWidget extends Widget implements HasForms
 {
-    use CanPoll;
     use CanDeferLoading;
     use CanFilter;
-    use HasHeader;
-    use HasFooter;
-    use HasLoadingIndicator;
-    use HasDarkMode;
+    use CanPoll;
     use HasContentHeight;
+    use HasDarkMode;
+    use HasFooter;
+    use HasHeader;
+    use HasLoadingIndicator;
 
     protected static string $chartId = 'apexChart';
 
@@ -90,5 +91,13 @@ class ApexChartWidget extends Widget implements HasForms
                     ->self();
             }
         }
+    }
+
+    /**
+     * Returns extra JavaScript options.
+     */
+    protected function extraJsOptions(): ?RawJs
+    {
+        return null;
     }
 }