+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 3.1.0
+
+### Added
+
+-   Extra options and Formatters.
+-   Funnel chart.
+
+### Changed
+
+-   Apexcharts update version.
+-   Commands.
+-   Plugin registration.
+
+## 3.0.2
+
+### Changed
+
+-   Apexcharts update version.
+-   Filament assets.
+
 ## 3.0.1
 
 ### Fixed